Refactoring wfs mvt and geojson implementations
