+++ bundles/mapping/mapwfs2/plugin/WfsVectorLayerPlugin.ol.js
@@ -1,90 +1,72 @@
-import olSourceVector from 'ol/source/Vector';
-import olLayerVector from 'ol/layer/Vector';
-import olFormatGeoJSON from 'ol/format/GeoJSON';
-import {bbox as bboxStrategy} from 'ol/loadingstrategy';
+import { VectorLayerHandler } from './WfsVectorLayerPlugin/impl/VectorLayerHandler.ol';
+import { MvtLayerHandler } from './WfsVectorLayerPlugin/impl/MvtLayerHandler.ol';
+import { ReqEventHandler } from './WfsVectorLayerPlugin/ReqEventHandler';
+import { HoverHandler } from './WfsVectorLayerPlugin/HoverHandler';
+import { styleGenerator } from './WfsVectorLayerPlugin/util/style';
+import { WFS_ID_KEY } from './WfsVectorLayerPlugin/util/props';
+import { LAYER_ID, LAYER_HOVER, LAYER_TYPE } from '../../mapmodule/domain/constants';
 
-import { selectedStyle } from './components/defaultStyle';
-import { styleGenerator, applyOpacity } from './components/styleUtils';
-import { WFS_ID_KEY } from './components/propertyArrayUtils';
-import {VectorPluginMixin} from './VectorPluginMixin.ol';
-import { LAYER_ID, LAYER_HOVER, LAYER_TYPE, FTR_PROPERTY_ID } from '../../mapmodule/domain/constants';
-
+const WfsLayerModelBuilder = Oskari.clazz.get('Oskari.mapframework.bundle.mapwfs2.domain.WfsLayerModelBuilder');
 const AbstractMapLayerPlugin = Oskari.clazz.get('Oskari.mapping.mapmodule.AbstractMapLayerPlugin');
-const MAP_MOVE_THROTTLE_MS = 2000;
-const OPACITY_THROTTLE_MS = 1500;
 
-class WfsVectorLayerPlugin extends VectorPluginMixin(AbstractMapLayerPlugin) {
+export class WfsVectorLayerPlugin extends AbstractMapLayerPlugin {
     constructor (config) {
         super(config);
         this.__name = 'WfsVectorLayerPlugin';
-        this._clazz = 'Oskari.wfsvector.WfsVectorLayerPlugin';
-        this.layertype = 'wfs';
-        this.hoverState = {
-            layer: null,
-            feature: null,
-            property: FTR_PROPERTY_ID
-        };
+        this._clazz = 'Oskari.wfs.WfsVectorLayerPlugin';
+        this.renderMode = config.renderMode || 'mvt';
+        this.visualizationForm = null;
+        this.oskariStyleSupport = true;
+        this.layertype = 'vector';
+        this.hoverHandler = new HoverHandler(WFS_ID_KEY);
+        this.vectorLayerHandler = new VectorLayerHandler(this);
+        this.mvtLayerHandler = new MvtLayerHandler(this);
+        this.layerHandlersByLayerId = {};
     }
-    /**
-     * @private @method _initImpl
-     * Interface method for the module protocol.
-     */
+
+    /* ---- AbstractMapModulePlugin functions ---- */
+
     _initImpl () {
         super._initImpl();
-        this.getSandbox().getService('Oskari.mapframework.service.VectorFeatureService')
-            .registerLayerType(this.layertype, this);
+        const sandbox = this.getSandbox();
+        // register domain builder
+        const mapLayerService = sandbox.getService('Oskari.mapframework.service.MapLayerService');
+        if (mapLayerService) {
+            mapLayerService.registerLayerModel(this.layertype + 'layer', this._getLayerModelClass());
+            mapLayerService.registerLayerModelBuilder(this.layertype + 'layer', this._getModelBuilder());
+        }
+        sandbox.getService('Oskari.mapframework.service.VectorFeatureService').registerLayerType(this.layertype, this);
+        this.WFSLayerService = Oskari.clazz.create(
+            'Oskari.mapframework.bundle.mapwfs2.service.WFSLayerService', sandbox);
+        sandbox.registerService(this.WFSLayerService);
+        this.reqEventHandler = new ReqEventHandler(sandbox);
+
+        this.visualizationForm = Oskari.clazz.create(
+            'Oskari.userinterface.component.VisualizationForm'
+        );
+    }
+    _getLayerModelClass () {
+        return 'Oskari.mapframework.bundle.mapwfs2.domain.WFSLayer';
+    }
+    _getModelBuilder () {
+        return new WfsLayerModelBuilder(this.getSandbox());
     }
     _createPluginEventHandlers () {
-        const updateStyle = event => this._updateLayerStyle(event.getMapLayer());
-
-        const throttleLoadFeatures = Oskari.util.throttle(
-            this._loadFeaturesForAllLayers, MAP_MOVE_THROTTLE_MS);
-
-        const throttleUpdateStyle = Oskari.util.throttle(
-            event => updateStyle(event), OPACITY_THROTTLE_MS);
-
         const handlers = {
-            ...super._createPluginEventHandlers(),
-            AfterMapMoveEvent: throttleLoadFeatures,
-            AfterChangeMapLayerStyleEvent: updateStyle,
-            AfterChangeMapLayerOpacityEvent: throttleUpdateStyle
+            // ...this.reqEventHandler.createEventHandlers(this),
+            ...this.vectorLayerHandler.createEventHandlers(this),
+            ...this.mvtLayerHandler.createEventHandlers(this)
         };
         return handlers;
     }
-    /**
-     * @private @method _updateLayerStyle
-     * @param {Oskari.mapframework.bundle.mapwfs2.domain.WFSLayer} oskariLayer
-     */
-    _updateLayerStyle (oskariLayer) {
-        if (!this.isLayerSupported(oskariLayer)) {
-            return;
-        }
-        const olLayers = this.getOLMapLayers(oskariLayer);
-
-        if (olLayers && olLayers.length > 0) {
-            const lyr = olLayers[0];
-            lyr.setStyle(this._getLayerCurrentStyleFunction(oskariLayer));
-            // Trigger features changed to synchronize 3D view
-            lyr.getSource().getFeatures().forEach(ftr => ftr.changed());
-        }
+    _createRequestHandlers () {
+        return this.reqEventHandler.createRequestHandlers(this);
     }
-    /**
-     * @private @method _getLayerCurrentStyleFunction
-     * Returns OL style corresponding to layer currently selected style
-     * @param {Oskari.mapframework.bundle.mapwfs2.domain.WFSLayer} layer
-     * @return {ol/style/Style}
-     */
-    _getLayerCurrentStyleFunction (layer) {
-        const factory = this.mapModule.getStyle.bind(this.mapModule);
-        const styleFunction = styleGenerator(factory, layer, this.hoverState);
-        const selectedIds = new Set(this.WFSLayerService.getSelectedFeatureIds(layer.getId()));
-
-        return (feature, resolution) => {
-            if (selectedIds.has(feature.get(WFS_ID_KEY))) {
-                return applyOpacity(selectedStyle(feature, resolution), layer.getOpacity());
-            }
-            return styleFunction(feature, resolution);
-        };
+    isLayerSupported (layer) {
+        if (!layer) {
+            return false;
+        }
+        return layer.isLayerOfType(this.layertype);
     }
     /**
      * @method addMapLayerToMap Adds wfs layer to map
@@ -93,107 +75,108 @@ class WfsVectorLayerPlugin extends VectorPluginMixin(AbstractMapLayerPlugin) {
      * @param {Boolean} isBaseMap
      */
     addMapLayerToMap (layer, keepLayerOnTop, isBaseMap) {
-        const source = this._getLayerSource(layer);
-        const vectorLayer = new olLayerVector({
-            source,
-            renderMode: 'image'
-        });
-        // Set oskari properties for vector feature service functionalities.
-        const silent = true;
-        vectorLayer.set(LAYER_ID, layer.getId(), silent);
-        vectorLayer.set(LAYER_TYPE, layer.getLayerType(), silent);
-        vectorLayer.set(LAYER_HOVER, layer.getHoverOptions(), silent);
-
-        this.mapModule.addLayer(vectorLayer, !keepLayerOnTop);
-        this.setOLMapLayers(layer.getId(), vectorLayer);
-        this._loadFeaturesForLayer(layer);
-    }
-
-    /**
-     * @private
-     * @method _getLayerSource To get an ol vector source for the layer.
-     * @param {Oskari.mapframework.bundle.mapwfs2.domain.WFSLayer} layer
-     * @return {ol.source.Vector} vector layer source
-     */
-    _getLayerSource (layer) {
-        var source = new olSourceVector({
-            format: new olFormatGeoJSON(),
-            url: Oskari.urls.getRoute('GetWFSFeatures'),
-            projection: this.getMap().getView().getProjection(), // OL projection object
-            strategy: bboxStrategy
-        });
-        source.setLoader(this._getFeatureLoader(layer, source));
-        return source;
-    }
-
-    /**
-     * @private
-     * @method _getFeatureLoader To get an ol loader impl for the layer.
-     * @param {Oskari.mapframework.bundle.mapwfs2.domain.WFSLayer} layer
-     * @param {ol.source.Vector} source
-     * @return {function} loader function for the layer
-     */
-    _getFeatureLoader (layer, source) {
-        return (extent, resolution, projection) => {
-            jQuery.ajax({
-                type: 'GET',
-                dataType: 'json',
-                data: {
-                    id: layer.getId(),
-                    srs: projection.getCode(),
-                    bbox: extent.join(',')
-                },
-                url: Oskari.urls.getRoute('GetWFSFeatures'),
-                success: (resp) => {
-                    source.addFeatures(source.getFormat().readFeatures(resp));
-                },
-                error: () => {
-                    source.removeLoadedExtent(extent);
-                }
-            });
-        };
-    }
-
-    /**
-     * @private
-     * @method _loadFeaturesForAllLayers Load features to all wfs layers.
-     * Uses current map view's extent.
-     */
-    _loadFeaturesForAllLayers () {
-        const mapView = this.mapModule.getMap().getView();
-        const extent = mapView.calculateExtent();
-        const resolution = mapView.getResolution();
-        const projection = mapView.getProjection();
-        this._sandbox.findAllSelectedMapLayers()
-            .filter(lyr => this.isLayerSupported(lyr))
-            .forEach(lyr => this._loadFeaturesForLayer(lyr, extent, resolution, projection));
-    }
-
-    /**
-     * @private
-     * @method _loadFeaturesForLayer Loads features to the layer.
-     *
-     * Uses OpenLayers internal methods.
-     * Load must be called manually in stacked 3D map mode.
-     * (No target container defined for 3D view)
-     *
-     * @param {Oskari.mapframework.bundle.mapwfs2.domain.WFSLayer} layer
-     * @param {ol.Extent} extent (optional)
-     * @param {Number} resolution (optional)
-     * @param {ol.proj.Projection} projection (optional)
-     */
-    _loadFeaturesForLayer (lyr, extent, resolution, projection) {
-        if (!extent) {
-            const mapView = this.mapModule.getMap().getView();
-            extent = mapView.calculateExtent();
-            resolution = mapView.getResolution();
-            projection = mapView.getProjection();
-        }
-        const olLayers = this.getOLMapLayers(lyr.getId());
-        if (olLayers.length === 0) {
+        const handler = this.renderMode === 'mvt' ? this.mvtLayerHandler
+            : this.renderMode === 'vector' ? this.vectorLayerHandler : null;
+        if (!handler) {
             return;
         }
-        olLayers[0].getSource().loadFeatures(extent, resolution, projection);
+        const added = handler.addMapLayerToMap(layer, keepLayerOnTop, isBaseMap);
+        if (!added) {
+            return;
+        }
+        this.layerHandlersByLayerId[added.getId()] = handler;
+        // Set oskari properties for vector feature service functionalities.
+        const silent = true;
+        added.set(LAYER_ID, layer.getId(), silent);
+        added.set(LAYER_TYPE, layer.getLayerType(), silent);
+        added.set(LAYER_HOVER, layer.getHoverOptions(), silent);
+
+        added.setStyle(handler.getCurrentStyleFunction(layer));
+    }
+
+    /* ----- VectorFeatureService interface functions ----- */
+
+    onMapHover (event, feature, layer) {
+        this.hoverHandler.onMapHover(event, feature, layer);
+    }
+    onLayerRequest (request, layer) {
+        this.hoverHandler.onLayerRequest(request, layer);
+    }
+
+    /* ---- Impl specific functions ---- */
+
+    /**
+     * @method getCustomStyleEditorForm To get editor ui element for custom style.
+     * @param {Oskari.mapframework.bundle.mapwfs2.domain.WFSLayer} layer
+     * @return VisualizationForm's form element
+     */
+    getCustomStyleEditorForm (layer) {
+        this.visualizationForm.setOskariStyleValues(layer.getCustomStyle());
+        return this.visualizationForm.getForm();
+    }
+    /**
+     * @method applyEditorStyle Applies custom style editor's style to the layer.
+     * @param {Oskari.mapframework.bundle.mapwfs2.domain.WFSLayer} layer
+     */
+    applyEditorStyle (layer) {
+        const style = this.visualizationForm.getOskariStyle();
+        layer.setCustomStyle(style);
+        layer.selectStyle('oskari_custom');
+    }
+    /**
+     * @method findLayerByOLLayer
+     * @param {ol/layer/Layer} olLayer OpenLayers layer impl
+     * @return {Oskari.mapframework.bundle.mapwfs2.domain.WFSLayer}
+     */
+    findLayerByOLLayer (olLayer) {
+        const layerId = Object.keys(this._layerImplRefs).find(layerId => olLayer === this._layerImplRefs[layerId]);
+        return this.getSandbox().getMap().getSelectedLayer(layerId);
+    }
+    /**
+     * @method getAllLayerIds
+     * @return {String[]} All layer ids handled by plugin and selected on map
+     */
+    getAllLayerIds () {
+        return Object.keys(this._layerImplRefs);
+    }
+
+    /**
+     * Helper to access the correct layer handler impl for a layer.
+     * @param {AbstractLayer | string | number} layer layer object or id
+     */
+    _getLayerHandler (layer) {
+        if (!layer) {
+            return;
+        }
+        const id = typeof layer === 'object' ? layer.getId() : layer;
+        return this.layerHandlersByLayerId[id];
+    }
+    /**
+     * @method getCurrentStyleFunction
+     * Returns OL style corresponding to layer currently selected style
+     * @param {Oskari.mapframework.bundle.mapwfs2.domain.WFSLayer} layer
+     * @return {ol/style/Style}
+     */
+    getCurrentStyleFunction (layer) {
+        const handler = this._getLayerHandler(layer);
+        if (!handler) {
+            return;
+        }
+        const factory = this.mapModule.getStyle.bind(this.mapModule);
+        const styleFunction = styleGenerator(factory, layer, this.hoverHandler);
+        return handler.getStyleFunction(layer, styleFunction);
+    }
+    /**
+     * @method updateLayerProperties
+     * Notify about changed features in view
+     * @param {Oskari.mapframework.bundle.mapwfs2.domain.WFSLayer} layer
+     * @param {ol/source/VectorTile} source
+     */
+    updateLayerProperties (layer, source) {
+        const handler = this._getLayerHandler(layer);
+        if (handler) {
+            return handler.updateLayerProperties(layer, source);
+        }
     }
 };
 
@@ -206,5 +189,3 @@ Oskari.clazz.defineES('Oskari.wfsvector.WfsVectorLayerPlugin', WfsVectorLayerPlu
         'protocol': ['Oskari.mapframework.module.Module', 'Oskari.mapframework.ui.module.common.mapmodule.Plugin']
     }
 );
-
-export {WfsVectorLayerPlugin};
