+++ bundles/mapping/mapwfs2/plugin/WfsVectorLayerPlugin/util/filter.js
