+++ bundles/mapping/mapwfs2/plugin/WfsVectorLayerPlugin/util/props.js
